More tests, more received types / response types Added several GameResponseTypes Added an interface to GameResponsesData for responses with a Chance in them (upgrade, compound) Added InviteData to received data types Updated several serialized collectioned to be IReadOnly collections Added some extension methods to AL.Client to help locate compoundable items Added a converter for Character to fix inventory size Added several new tests to cover gaps in deserialization

+++ AL.SocketClient/SocketModel/PartyMember.cs
@@ -15,7 +15,7 @@ namespace AL.SocketClient.SocketModel
         public string Map { get; init; }
         [JsonProperty("l")]
         public int PartyLimit { get; init; }
-        public int Share { get; init; }
+        public float Share { get; init; }
         public string Skin { get; init; }
         public float X { get; init; }
         public int XP { get; init; }
